Clean up and tweaks
